(#issue-28)
* Разделение API аппа Operations на отдельные файлы
