+++ server/app/operations/apis/__init__.py
@@ -1,0 +1,11 @@
+
+from .category import CategoryConcreteView, CategoryGeneralView
+from .operation import OperationConcreteView, OperationGeneralView
+
+
+__all__ = [
+    "CategoryConcreteView",
+    "CategoryGeneralView",
+    "OperationConcreteView",
+    "OperationGeneralView",
+]
